enable warnings\nEnable warnings for compile on linux; fix all the warnings identified

+++ src/bwm801_calllist.cpp
@@ -4,11 +4,11 @@ using namespace bwm801;
 
 bool calllist::Is_Valid(void) const
 {
-	return (m_uiList_ID != -1);
+	return (m_uiList_ID != (unsigned int)-1);
 }
 void calllist::Compile(bool i_bCompile_and_Exec)
 {
-	if (m_uiList_ID == -1)
+	if (m_uiList_ID == (unsigned int)-1)
 		m_uiList_ID = glGenLists(1);
 #ifdef _DEBUG
 	_RPT1(_CRT_WARN, "Call list created: %i\n", m_uiList_ID);
@@ -21,19 +21,19 @@ void calllist::End_Compile(void)
 }
 void calllist::Draw(void)
 {
-	if (m_uiList_ID != -1)
+	if (m_uiList_ID != (unsigned int)-1)
 		glCallList(m_uiList_ID);
 }
 void	calllist::Delete(void)
 {
-	if (m_uiList_ID != -1)
+	if (m_uiList_ID != (unsigned int)-1)
 		glDeleteLists(m_uiList_ID, 1);
-	m_uiList_ID = -1;
+	m_uiList_ID = (unsigned int)-1;
 }
 
 void calllist::Initializer(void)
 {
-	m_uiList_ID = -1;
+	m_uiList_ID = (unsigned int)-1;
 }
 void calllist::Destructor(void)
 {
